move all dev pages to /dev
